Continued to implement reminder related functionality.
Added two methods to ChannelAPI: changeActiveStatusOfReminder and changeIgnoreFlagOfReminder.
Removed the ignore flag related functionality from the CreateOrEditReminderAction and placed it seperately in RemindersAction.
Added functionality for activating/deactivating reminder and set/reset of ignore flag in RemindersAction.
Added modal dialogs and buttons for this kind of functionality to the reminders.jsp view. Added icons to represent the status of a reminder.

Changed icons in manageChannelModerators.jsp.