Changed some date related things. This was done especially for reminder handling.
Set default time zone to Europe/Berlin in Constants. Dates in web client should always be considered to be defined relating to this time zone.

Adjusted method retrieveLocale() in RequestContextManager to fix some bugs. A problem occurred when reloading the reminder pages choosing another language using the nav bar. The new locale was set in the jsp, but the servlets are executed before the jsp is loaded. Adjusted the method to take the language parameter into account if it is set. This fixes the problem.

Adjusted createReminder.jsp and editReminder.jsp. Fixed some issue that appeared in Edge browser. Locale handling is a problem if the browser sets the locale using the accept parameter (or language paramere) from the request. In edge, this is de_DE instead of just de. Added some additional cases to represent this. Fallback is to choose english as default language.
Furthermore set autocomplete to off for date related input fields.

The locale related problems were also considered in the action instances. Changed implementation of CreateOrEditReminderAction and LoadEditReminderAction. Now checks for different locales, e.g. Locale.GERMAN and Locale.GERMANY. Added default case with english language.